+++ src/main/java/page/PageException.java
@@ -1,3 +1,5 @@
+package page;
+
 public class PageException extends Exception {
     public PageException(String errorMessage) {
         super(errorMessage);
